movie_agent_project/requirements.txt
weaviate-client[agents]>=4.9.4
streamlit
python-dotenv
pandas
